answers for lesson 1.9
