+++ src/App.js
@@ -1,117 +1,15 @@
-import React, { useEffect, useState } from 'react';
-import AddTodoForm from './AddTodoForm';
-import TodoList from './TodoList';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import MainFile from './MainFile';
 
-const App = () => {
-  const [todoList, setTodoList] = useState();
-  const [isLoading, setIsLoading] = useState(true);
-
-  const fetchData = async () => {
-    const url = `https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/${process.env.REACT_APP_TABLE_NAME}`;
-    const options = {
-      method: "GET",
-      headers: {
-        Authorization: `Bearer ${process.env.REACT_APP_AIRTABLE_API_KEY}`,
-      },
-    };
-    
-    try {
-      const response = await fetch(url, options);
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-      const data = await response.json();
-      const todos = data.records.map((todo) => {
-        return {
-          id: todo.id,
-          title: todo.fields.Title
-        };
-      });
-
-      setTodoList(todos);
-      setIsLoading(false);
-
-    } catch (error) {
-      console.log(error.message);
-    }
-  };  
-
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  const addTodo = async (title) => {
-    const url = `https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/${process.env.REACT_APP_TABLE_NAME}`;
-    const newTitle = {
-      fields: {
-        Title: title,
-      },
-    };
-    const options = {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${process.env.REACT_APP_AIRTABLE_API_KEY}`,
-        "Content-type": "application/json",
-      }, 
-      body: JSON.stringify(newTitle),
-    };
-
-    try {
-      const response = await fetch(url, options);
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      const todo = await response.json();
-      const newTodo = {
-          id: todo.id,
-          title: todo.fields.Title
-        }; 
-
-      console.log(newTodo);
-      setTodoList([...todoList, newTodo]);
-
-    } catch (error) {
-      console.log(error.message);
-    }
-  };
-
-  const removeTodo = async(id) => {
-    const url = `https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/${process.env.REACT_APP_TABLE_NAME}/${id}`;
-    const options = {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${process.env.REACT_APP_AIRTABLE_API_KEY}`,
-        "Content-type": "application/json",
-      },
-    };
-    try {
-      const response = await fetch(url, options);
-        if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-      const newTodoList = todoList.filter(
-        item => item.id !== id);
-      setTodoList(newTodoList);
-    } catch (error) {
-      console.log(error.message);
-    }
-  };
-
+function App() {
   return (
-    <>
-      <h1>To Do List</h1>
-
-      <AddTodoForm onAddTodo={addTodo} />
-      {isLoading ? (
-          <p>Loading...</p>
-        ): (
-            <TodoList todoList={todoList} onRemoveTodo={removeTodo}/>
-      )}
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route exact path="/" element={<MainFile/>} />
+        <Route exact path="/new" element={<h1>New Todo List</h1>} />
+      </Routes>
+    </BrowserRouter>
   );
-}
+};
 
 export default App;
